add project switcher as independent component
